requirement.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/croot-2p6q7e2r/cffi_1642701115063/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
coreapi==2.3.3
coreschema==0.0.4
cryptography @ file:///opt/concourse/worker/volumes/live/732d2752-e4cb-4715-726b-56d890f169fb/volume/cryptography_1639414586414/work
Django==4.0.2
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2-binary==2.9.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///opt/concourse/worker/volumes/live/85a5b906-0e08-41d9-6f59-084cee4e9492/volume/pysocks_1594394636991/work
pytz==2021.3
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
sqlparse==0.4.2
uritemplate==4.1.1
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
